requirements.txt
PyMuPDF==1.18.19
google-generativeai==0.1.0
python-dotenv==0.19.1